created new component TodoTask and transferred logic inside
