+++ src/index.css
@@ -168,7 +168,7 @@ body {
   background: #5292fe;
 }
 
-.main__task_buttons-cansel {
+.main__task_buttons-canсel {
   background: #b3ada9;
 }
